% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_hua_clean.R
\name{linear_regression}
\alias{linear_regression}
\title{A custom function that calculates multiple linear regress}
\usage{
linear_regression(predictor, outcome)
}
\arguments{
\item{predictor}{the predictor variables}

\item{outcome}{the outcome variables}
}
\value{
returns beta, SE, t value, p value of the linear regression
}
\description{
It returns a table of beta coefficient, standard error, t value, and p value
as variables, and predictors as observations,
}
\examples{
Y = c(5.6, 7.9, 10.8)
X = c(1, 2, 3)
linear_regression(X, Y)

}
